backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
python-dotenv==1.0.0
gunicorn==21.2.0
PyJWT==2.8.0
setuptools==78.1.0
whitenoise==6.6.0